Spring Balance Sheet
This is a simple application using the Spring MVC Framework in Java.
This application is a simple balance sheet used in accounting, The user
will calculate the balance sheet, and use the Hibernate ORM to map
persistent objects to a oracle database.

The Spring and Hibernate uses Contextual Sessions to integrate the two
frameworks together,